Aerial Object Classification & Detection/src/evaluation/confusion_matrix_plot.py
from __future__ import annotations

from pathlib import Path
from typing import Sequence

import matplotlib.pyplot as plt
import numpy as np
from sklearn.metrics import ConfusionMatrixDisplay


def save_confusion_matrix_figure(
    confusion_matrix: Sequence[Sequence[int]],
    class_names: Sequence[str],
    output_path: Path,
    title: str,
    dpi: int = 150,
) -> Path:
    """
    Save a confusion matrix figure and return the output path.
    """
    output_path = Path(output_path)
    output_path.parent.mkdir(parents=True, exist_ok=True)

    cm = np.array(confusion_matrix)

    fig, ax = plt.subplots(figsize=(6, 5))
    disp = ConfusionMatrixDisplay(confusion_matrix=cm, display_labels=list(class_names))
    disp.plot(ax=ax, cmap="Blues", colorbar=False)
    ax.set_title(title)

    plt.tight_layout()
    plt.savefig(output_path, dpi=dpi, bbox_inches="tight")
    plt.close()

    return output_path