Aerial Object Classification & Detection/requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
torch
torchvision
ultralytics
opencv-python
pillow
pyyaml
streamlit
tqdm
seaborn